Pull functions into util module.
